+++ packages/core/src/default-compiler-options.ts
@@ -3,12 +3,28 @@ import ts from 'typescript';
 
 /**
  * These are the default TypeScript Compiler options used when we're unable
- * to find the `tsconfig.json` in the root of the project.
+ * to find TSConfig file in the root of the project.
  */
-export const DEFAULT_COMPILER_OPTIONS: ts.CompilerOptions = {
+export const TS_DEFAULT_COMPILER_OPTIONS: ts.CompilerOptions = {
     experimentalDecorators: true,
     target: ts.ScriptTarget.Latest,
     module: ts.ModuleKind.ESNext,
+    skipLibCheck: true,
+    skipDefaultLibCheck: true,
     declaration: true,
+};
+
+/**
+ * These are the default TypeScript Compiler options used when we're analyzing
+ * a JavaScript project.
+ */
+export const JS_DEFAULT_COMPILER_OPTIONS: ts.CompilerOptions = {
+    module: ts.ModuleKind.ESNext,
+    lib: ['es2020', 'DOM'],
     allowJs: true,
+    skipLibCheck: true,
+    skipDefaultLibCheck: true,
+    target: ts.ScriptTarget.Latest,
+    moduleResolution: ts.ModuleResolutionKind.NodeNext,
+    typeRoots: [],
 };