+++ packages/core/src/nodes/namespace-import-node.ts
@@ -49,7 +49,7 @@ export class NamespaceImportNode implements ReflectedNode<Import, ts.ImportDecla
         const identifier = (this._node.importClause?.namedBindings as ts.NamespaceImport)?.name;
         const importPath = this.getImportPath();
 
-        return matchesTsConfigPath(importPath, this._context.compilerOptions)
+        return matchesTsConfigPath(importPath, this._context.options?.compilerOptions ?? {})
             ? getOriginalImportPath(identifier, this._context)
             : importPath;
     }