+++ packages/core/src/resolve-compiler-options.ts
@@ -1,23 +1,77 @@
-import { DEFAULT_COMPILER_OPTIONS } from './default-compiler-options.js';
+import { JS_DEFAULT_COMPILER_OPTIONS, TS_DEFAULT_COMPILER_OPTIONS } from './default-compiler-options.js';
+import type { AnalyzerOptions } from './analyzer-options.js';
 import ts from 'typescript';
 import path from 'path';
 
 
-export function getResolvedCompilerOptions(compilerOptions?: ts.CompilerOptions): ts.CompilerOptions {
-    if (!compilerOptions) {
-        return getTsConfigOptions() ?? DEFAULT_COMPILER_OPTIONS;
+export interface ResolvedCompilerOptions {
+    compilerOptions: ts.CompilerOptions;
+    commandLine: ts.ParsedCommandLine | null;
+}
+
+export function getResolvedCompilerOptions(options?: Partial<AnalyzerOptions>): ResolvedCompilerOptions {
+    const {compilerOptions, jsProject, tsConfigFilePath} = options ?? {};
+
+    // If it's a JS project, we currently don't allow the user to customize the compiler options
+    if (jsProject) {
+        return {
+            compilerOptions: JS_DEFAULT_COMPILER_OPTIONS,
+            commandLine: ts.parseJsonConfigFileContent(
+                {
+                    compilerOptions: JS_DEFAULT_COMPILER_OPTIONS,
+                    include: ['**/*.js'],
+                    exclude: ['**node_modules**'],
+                },
+                ts.sys,
+                process.cwd(),
+            ),
+        };
+    }
+
+    // If it's a TS project and the user provides us it's custom compiler options, we will use them
+    if (compilerOptions) {
+        return {
+            compilerOptions: {...compilerOptions, declaration: true},
+            commandLine: ts.parseJsonConfigFileContent(
+                {
+                    compilerOptions: {...compilerOptions, declaration: true},
+                    include: ['**/*.ts'],
+                    exclude: ['**node_modules**'],
+                },
+                ts.sys,
+                process.cwd(),
+            ),
+        };
+    }
+
+    // If user doesn't provide the compiler options, we will resolve them by
+    // searching for a TSConfig file
+    const parsedOptions = parseTSConfigFile(tsConfigFilePath);
+
+    if (parsedOptions) {
+        return parsedOptions;
     }
 
     return {
-        ...compilerOptions,
-        declaration: true,
+        compilerOptions: TS_DEFAULT_COMPILER_OPTIONS,
+        commandLine: ts.parseJsonConfigFileContent(
+            {
+                compilerOptions: TS_DEFAULT_COMPILER_OPTIONS,
+                include: ['**/*.ts'],
+                exclude: ['**node_modules**'],
+            },
+            ts.sys,
+            process.cwd(),
+        ),
     };
 }
 
-function getTsConfigOptions(): ts.CompilerOptions | null {
+function parseTSConfigFile(tsConfigFilePath?: string): ResolvedCompilerOptions | null {
     const fileExists = (filePath: string) => ts.sys.fileExists(filePath);
     const readFile = (filePath: string) => ts.sys.readFile(filePath);
-    const configFileName = ts.findConfigFile(process.cwd(), fileExists, 'tsconfig.json');
+    const basePath = tsConfigFilePath ? path.dirname(path.join(process.cwd(), tsConfigFilePath)) : process.cwd();
+    const fileName = tsConfigFilePath ? path.basename(tsConfigFilePath) : 'tsconfig.js';
+    const configFileName = ts.findConfigFile(basePath, fileExists, fileName);
     const configFile = configFileName && ts.readConfigFile(configFileName, readFile);
 
     if (configFile && typeof configFile === 'object') {
@@ -30,8 +84,8 @@ function getTsConfigOptions(): ts.CompilerOptions | null {
         );
 
         return {
-            ...commandLine.options,
-            declaration: true,
+            commandLine,
+            compilerOptions: {...commandLine.options, declaration: true},
         };
     }